requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
pydantic==2.6.0
python-dotenv==1.0.0

# Machine Learning
xgboost==2.0.3
tensorflow==2.15.0
scikit-learn==1.4.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# HTTP & Scraping
requests==2.31.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
httpx==0.26.0

# LLM Integration
openai==1.12.0

# Visualization
plotly==5.18.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Database
aiosqlite==0.19.0

# Utilities
schedule==1.2.1
pydantic-settings==2.1.0
